Fix up DNS cleanup to deploy
