+++ tests/test_dns_cleanup.py
@@ -6,7 +6,7 @@ import os
 from unittest.mock import MagicMock, patch
 
 from devbox.lifecycle.dns import cleanup_dns as lifecycle_cleanup_dns
-from lambdas.dns_cleanup import cleanup_dns as lambda_cleanup_dns
+from devbox.lambdas.dns_cleanup import cleanup_dns as lambda_cleanup_dns
 
 
 class TestLifecycleCleanupDns:
@@ -97,10 +97,10 @@ class TestLifecycleCleanupDns:
 
 
 class TestLambdaCleanupDnsWrapper:
-    @patch("lambdas.dns_cleanup.dns_lifecycle.cleanup_dns")
-    @patch("lambdas.dns_cleanup.utils.get_ssm_client")
-    @patch("lambdas.dns_cleanup.utils.get_ec2_client")
-    @patch("lambdas.dns_cleanup.utils.get_dynamodb_table")
+    @patch("devbox.lambdas.dns_cleanup.dns_lifecycle.cleanup_dns")
+    @patch("devbox.lambdas.dns_cleanup.utils.get_ssm_client")
+    @patch("devbox.lambdas.dns_cleanup.utils.get_ec2_client")
+    @patch("devbox.lambdas.dns_cleanup.utils.get_dynamodb_table")
     def test_lambda_wrapper_delegates_to_lifecycle(
         self,
         mock_get_table,
